+++ src/main/java/net/rmoff/connect/ais/AisSourceTask.java
@@ -31,7 +31,7 @@ public class AisSourceTask extends SourceTask {
 
     @Override
     public String version() {
-        return "0.2.0";
+        return "0.2.1";
     }
 
     @Override